==> Part2/Part2/1_GlobalFeatureExtraction.ipynb <==
### Import Libraries

# supresses warning messages
import warnings
warnings.filterwarnings("ignore")

import pandas as pd # for data manipulation and analysis
import numpy as np # numerical computations 
import hashlib # secure hash and message digest algorithms
import os # interacting with operating sys

# reads the list of filenames (i.e., video labels) in the "data" directory 
# and saves them into a Python list called "videos_labels".
videos_labels=os.listdir("data")
print(videos_labels)

df=pd.DataFrame()
df["Video Names"]=videos_labels

# calculates SHA256 hashes of each video file in the "data" directory and saves the hashes in a list called "Features". 
BLOCK_SIZE = 65536
Features=[]
for i in range(len(videos_labels)):
    video=videos_labels[i]
    filepath=os.path.join("data",video)
    file_hash = hashlib.sha256()
    with open(filepath, 'rb') as f: 
        fb = f.read(BLOCK_SIZE)
        while len(fb) > 0:
            file_hash.update(fb)
            fb = f.read(BLOCK_SIZE)
    Features.append(file_hash.hexdigest())

print(Features)

df["features"]=Features

df.head()

df.to_csv("Feature_Files/GlobalFeatureExtraction.csv",index=False)

## Thank You
