==> Part2/Part2/2_LocalFeatureExtraction.ipynb <==
### Import Libraries

import cv2 # for image processing
import pandas as pd # data manipulation
import numpy as np # numerical computation
from glob import glob #  find all pathnames matching a specified pattern in a single line
import hashlib # secure hash and message digest algorithms
import os
from tensorflow.keras.preprocessing.image import load_img,img_to_array # for buildong CNN Model
from tensorflow.keras.models import model_from_json
from tensorflow.keras import Model
from pathlib import Path # for working on paths

# creates a list called file_path which contains the full file path of each file in the "data" directory.
import os
file_path=[]
files=os.listdir("data")
for i in range(len(files)):
    file_path.append(os.path.join("data",files[i]))

print(file_path)

# reads in the video files in the file_path list, and extracts every frame from each video 
# and saves them as individual image files in a directory called "Extracted_Frames".
count = 0

for i in range(0,6):
    cam = cv2.VideoCapture(file_path[i])
    currentframe = 0
    while(True): 
        ret,frame = cam.read() 
        if ret:
            if count%1 == 0 :
                filename = 'Extracted_Frames/'+file_path[i].split('\\')[1]+"_frame%d.jpg" % count;
                print ('Creating...' + filename) 
                cv2.imwrite(filename, frame) 
                currentframe += 1
            count+=1
        else: 
            break
    cam.release() 
    #cv2.destroyAllWindows() 

file = Path("cnn_model/model_structure.json")
model_structure = file.read_text()
model = model_from_json(model_structure)
model.load_weights("cnn_model/model_weights.h5")

# prints the name of each layer in a pre-trained model.
for idx in range(len(model.layers)):
    print(model.get_layer(index = idx).name)

# for feature extraction we are going till dense layer
intermediate_layer_model = Model(inputs=model.input,outputs=model.get_layer('dense').output)

# finding all the JPEG image files in the "Extracted_Frames" directory and 
# printing the count of the total number of such images found.
images = glob("Extracted_Frames/*.jpg")
print("Number of frames:{}".format(len(images)))

print(images)

# reads in the extracted frames, resizes them to 64x64 pixels, and normalizes their pixel values between 0 and 1.
train_image = []
frame_names = []
video_name = []
for i in range(len(images)):
    img = load_img(images[i], target_size=(64,64,3))
    img = img_to_array(img)
    img = img/255.0
    train_image.append(img)
    f_name=images[i]
    f_names=f_name.split("\\")[1]
    f_names=f_names.split(".jpg")[0]
    f_names=f_names.split(".mp4_")[1]
    frame_names.append(f_names)
    f1_names=f_name.split("\\")[1]
    f1_names=f1_names.split(".mp4_")[0]
    video_name.append(f1_names)
X = np.array(train_image)

# to predict the output values for the given input data X, which consists of a set of images. 
intermediate_output = intermediate_layer_model.predict(X)

df=pd.DataFrame(intermediate_output)

df.shape

column_names=[]
for i in range(1024):
    column_names.append("fet_"+str(i))

df.columns=column_names

df.insert(loc=0,column="VideosNames",value=video_name)

df.insert(loc=1,column="FramesNames",value=frame_names)

df.tail()

df.to_csv("Feature_Files/LocalFeatureExtraction.csv",index=False)

## Thank You
