==> Part2/Part2/3_Test_File_Ref.ipynb <==
# supresses warning messages
import warnings
warnings.filterwarnings("ignore")

import pandas as pd # data manipulation
import numpy as np # numerical computation
import cv2 # for image processing
import hashlib # secure hash and message digest algorithms
import os # file manipulation
from glob import glob #  find all pathnames matching a specified pattern in a single line
from tensorflow.keras.preprocessing.image import load_img,img_to_array # for buildong CNN Model
from tensorflow.keras.models import model_from_json
from tensorflow.keras import Model
from pathlib import Path # for working on paths

os.listdir("Feature_Files")

global_df=pd.read_csv("Feature_Files/GlobalFeatureExtraction.csv")
local_df=pd.read_csv("Feature_Files/LocalFeatureExtraction.csv")

global_df.head()

local_df.head()

file = Path("cnn_model/model_structure.json")
model_structure = file.read_text()
model = model_from_json(model_structure)
model.load_weights("cnn_model/model_weights.h5")

for idx in range(len(model.layers)):
    print(model.get_layer(index = idx).name)

# for feature extraction we are going till dense layer
intermediate_layer_model = Model(inputs=model.input,outputs=model.get_layer('dense').output)

### UserInput

#user_input="UserInput/y2.mp4" # case 1 : existing video-> options: (y1.mp4,y2.mp4)
#user_input="UserInput/n8.mp4" # case 2: partial match-> options: (n8.mp4)
user_input="UserInput/n3.mp4" # case 3: new video-> options: (n3.mp4,n5.mp4)

# calculates SHA256 hashes of each video file in the "data" directory and saves the hashes in a list called "Features". 
BLOCK_SIZE = 65536
Features = []
file_hash = hashlib.sha256()
with open(user_input, 'rb') as f: 
    fb = f.read(BLOCK_SIZE)
    while len(fb) > 0:
        file_hash.update(fb)
        fb = f.read(BLOCK_SIZE)
Features.append(file_hash.hexdigest())

print(Features)

matched_videoName=[]
for i in range(len(Features)):
    for j in range(global_df.shape[0]):
        if str(global_df["features"][j])==str(Features[i]):
            matched_videoName.append(global_df["Video Names"][j])
if len(matched_videoName)>0:
    print("Matched video name is {}".format(matched_videoName[0]))
else:
    print("There is no matching")

---

### Delete old files

images = glob("UserInput_Frames/*.jpg")
for old_images in images:
    os.remove(old_images)

### Converting video into frames

count = 0
cam = cv2.VideoCapture(user_input)
while(True): 
    ret,frame = cam.read() 
    if ret:
        filename = 'UserInput_Frames/'+user_input.split('/')[1]+"_frame%d.jpg" % count;
        print ('Creating...' + filename) 
        cv2.imwrite(filename, frame) 
        count+=1
    else: 
        break
cam.release() 

### Finding LOCAL FEATURES for all frames

input_frames_names=glob("UserInput_Frames/*.jpg")

print(input_frames_names)

train_image = []
frame_name = []
for i in range(len(input_frames_names)):
    img = load_img(input_frames_names[i], target_size=(64,64,3))
    img = img_to_array(img)
    img = img/255.0
    train_image.append(img)
    f_name=input_frames_names[i]
    f_name=f_name.split("\\")[1]
    f_name=f_name.split("_")[1]
    frame_name.append(f_name)
X = np.array(train_image)

print(frame_name)

intermediate_output = intermediate_layer_model.predict(X)

intermediate_output.shape

df=pd.DataFrame(intermediate_output)

column_names=[]
for i in range(1024):
    column_names.append("pxl_"+str(i))

df.columns=column_names

df.insert(loc=0,column="FrameName",value=frame_name)

df.head()

local_df.head()

local_df1=local_df.drop(labels=["VideosNames","FramesNames"],axis=1)

local_df1.head()

local_df1.shape

df1=df.drop(labels="FrameName",axis=1)

df1.head()

df1.shape

local_df1.shape

    # Euclidean distance finding
    loc_ix=[]
    for i in range(len(df1)):
        for j in range(len(local_df1)):
            similarity = np.dot(local_df1.iloc[j].values, df1.iloc[i].values) / (np.linalg.norm(local_df1.iloc[j].values) * np.linalg.norm(df1.iloc[i].values))
            if similarity>0.90:
                loc_ix.append(j)
            else:
                continue

print(loc_ix)

loc_ix=list(set(loc_ix))

print(loc_ix)

### Duplicated Frames form the input data

duplicated_frames=local_df.iloc[loc_ix]

duplicated_frames.shape

duplicated_frames=duplicated_frames[['VideosNames','FramesNames']]
duplicated_frames.head()



# Thank You
